refactor: move all link-format files into one folder
