+++ lib/coap.dart
@@ -22,7 +22,6 @@ export 'src/coap_code.dart';
 export 'src/coap_config.dart';
 export 'src/coap_constants.dart';
 export 'src/coap_defined_address.dart';
-export 'src/coap_link_format.dart';
 export 'src/coap_media_type.dart';
 export 'src/coap_message_type.dart';
 export 'src/coap_observe_client_relation.dart';
@@ -35,5 +34,6 @@ export 'src/deduplication/coap_ideduplicator.dart';
 export 'src/deduplication/coap_noop_deduplicator.dart';
 export 'src/deduplication/coap_sweep_deduplicator.dart';
 export 'src/exceptions/coap_request_exception.dart';
+export 'src/link-format/coap_link_format.dart';
 export 'src/network/credentials/ecdsa_keys.dart';
 export 'src/network/credentials/psk_credentials.dart';
