+++ lib/src/coap_client.dart
@@ -17,7 +17,6 @@ import 'coap_code.dart';
 import 'coap_config.dart';
 import 'coap_constants.dart';
 import 'coap_empty_message.dart';
-import 'coap_link_format.dart';
 import 'coap_media_type.dart';
 import 'coap_message.dart';
 import 'coap_message_type.dart';
@@ -26,9 +25,10 @@ import 'coap_option.dart';
 import 'coap_option_type.dart';
 import 'coap_request.dart';
 import 'coap_response.dart';
-import 'coap_web_link.dart';
 import 'event/coap_event_bus.dart';
 import 'exceptions/coap_request_exception.dart';
+import 'link-format/coap_link_format.dart';
+import 'link-format/coap_web_link.dart';
 import 'net/coap_endpoint.dart';
 import 'net/coap_iendpoint.dart';
 import 'network/coap_inetwork.dart';
