+++ lib/src/link-format/resources/coap_resource_attributes.dart
